+++ src/main/java/com/jnape/palatable/lambda/functions/Fn2.java
@@ -3,6 +3,7 @@ package com.jnape.palatable.lambda.functions;
 import com.jnape.palatable.lambda.adt.hlist.Tuple2;
 
 import java.util.function.BiFunction;
+import java.util.function.Function;
 
 /**
  * A function taking two arguments. Note that defining <code>Fn2</code> in terms of <code>Fn1</code> provides a
@@ -25,6 +26,18 @@ public interface Fn2<A, B, C> extends Fn1<A, Fn1<B, C>> {
      * @return the result of the function application
      */
     C apply(A a, B b);
+
+    /**
+     * Same as normal composition, except that the result is an instance of <code>Fn2</code> for convenience.
+     *
+     * @param before the function who's return value is this function's argument
+     * @param <Z>    the new argument type
+     * @return a new Fn2&lt;Z,B,C&gt;
+     */
+    @Override
+    default <Z> Fn2<Z, B, C> compose(Function<? super Z, ? extends A> before) {
+        return fn2(Fn1.super.compose(before));
+    }
 
     /**
      * Partially apply this function by passing its first argument.
@@ -74,8 +87,37 @@ public interface Fn2<A, B, C> extends Fn1<A, Fn1<B, C>> {
      * @param <B>        the second input argument type
      * @param <C>        the output type
      * @return the Fn2
+     * @deprecated in favor of {@link Fn2#fn2(BiFunction)}
      */
+    @Deprecated
     static <A, B, C> Fn2<A, B, C> adapt(BiFunction<A, B, C> biFunction) {
         return biFunction::apply;
+    }
+
+    /**
+     * Static factory method for wrapping a {@link BiFunction} in an {@link Fn2}. Useful for avoid explicit casting when
+     * using method references as {@link Fn2}s.
+     *
+     * @param biFunction the biFunction to adapt
+     * @param <A>        the first input argument type
+     * @param <B>        the second input argument type
+     * @param <C>        the output type
+     * @return the Fn2
+     */
+    static <A, B, C> Fn2<A, B, C> fn2(BiFunction<A, B, C> biFunction) {
+        return biFunction::apply;
+    }
+
+    /**
+     * Static factory method for wrapping a curried {@link Fn1} in an {@link Fn2}.
+     *
+     * @param curriedFn1 the curried fn1 to adapt
+     * @param <A>        the first input argument type
+     * @param <B>        the second input argument type
+     * @param <C>        the output type
+     * @return the Fn2
+     */
+    static <A, B, C> Fn2<A, B, C> fn2(Fn1<A, Fn1<B, C>> curriedFn1) {
+        return (a, b) -> curriedFn1.apply(a).apply(b);
     }
 }